Create AllMolds mold to show contents of me-available-molds
This uses the same format as WhatMoldsCanIUse?
Because it is a direct copy, using the full list, rather than the
filtered one.

This helps mold writers get access to existing molds for inspiration,
and to see what is available.

It might be interesting to add a column for the major mode each is
applicable in.

I put it in contrib, no idea where it really belongs.